implemented grandparent+uncle getters
Implemented the functions which return the current node's grandparent and uncle. Information about the grandparent and uncle nodes is critical for maintaining balance in a red-black tree. 

Finding the grandparent was pretty easy, since we just need to return the current node's parent's parent. Finding the uncle is a bit more involved, since we need to determine whether the node's parent is the left or right child of the grandparent.