Added new tests
I have added new tests to verify that the insert function behaves as it should. I was a bit confused since I assumed the tree would be perfectly balanced when it contains 1-7, but after drawing the state of the tree I now understand that even when not perfectly balanced it still follows the rules of the red-black tree.