Declared remove cases
I am having each case of a remove be handled by its own function. In these functions the pointers will be properly adjusted according to the case. I am doing it this way instead of having it all be done in the normal removeNode() function to help make code more legible and debug easier.